Функции по отправки сообщений и попыток рассылки вынесены в модуль services.py
